+++ Android-app/ParkWire/app/src/main/java/parkwire/com/models/Paid.java
@@ -5,7 +5,7 @@ public class Paid extends Parking{
     private int capacity;
     private float cost;
     private String info;
-    public Paid(float lat, float lon, Boolean free, Boolean meter, int cap, float cost, String info){
+    public Paid(double lat, double lon, Boolean free, Boolean meter, int cap, float cost, String info){
         super(lat, lon, free, meter);
         this.capacity = cap;
         this.cost = cost;
