+++ Android-app/ParkWire/app/src/main/java/parkwire/com/models/Parking.java
@@ -1,12 +1,12 @@
 package parkwire.com.models;
 
 public class Parking {
-    private float lat;
-    private float lon;
+    private double lat;
+    private double lon;
     private Boolean isFree;
     private Boolean isMeter;
 
-    public Parking(float lat, float lon, Boolean free, Boolean meter){
+    public Parking(double lat, double lon, Boolean free, Boolean meter){
         this.lat = lat;
         this.lon = lon;
         this.isFree = free;
